regulatory-intelligence-crawler/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Web scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Vertex AI
google-cloud-aiplatform==1.40.0
google-generativeai==0.3.2

# Database
google-cloud-firestore==2.14.0

# Utilities
python-dateutil==2.8.2
